Split AggregrateType out of type.{h,cpp} and in to AggregateType.{h,cpp}
