+++ compiler/AST/Makefile.share
@@ -16,6 +16,7 @@
 # limitations under the License.
 
 AST_SRCS =                                          \
+           AggregateType.cpp                        \
            alist.cpp                                \
            astutil.cpp                              \
            baseAST.cpp                              \
