+++ compiler/include/type.h
@@ -346,147 +346,7 @@ private:
 *                                                                             *
 ************************************** | *************************************/
 
-enum AggregateTag {
-  AGGREGATE_CLASS,
-  AGGREGATE_RECORD,
-  AGGREGATE_UNION
-};
-
-enum InitializerStyle {
-  DEFINES_CONSTRUCTOR,
-  DEFINES_INITIALIZER,
-  DEFINES_NONE_USE_DEFAULT
-};
-
-
-class AggregateType : public Type {
-public:
-  static void                 addRecordDefaultConstruction();
-
-  static void                 setCreationStyle(TypeSymbol* t, FnSymbol* fn);
-
-public:
-                              AggregateType(AggregateTag initTag);
-                             ~AggregateType();
-
-  DECLARE_COPY(AggregateType);
-
-  virtual void                replaceChild(BaseAST* oldAst, BaseAST* newAst);
-
-  virtual void                verify();
-
-  virtual void                accept(AstVisitor* visitor);
-
-  virtual void                printDocs(std::ostream* file, unsigned int tabs);
-
-  void                        addDeclarations(Expr* expr);
-
-  void                        codegenDef();
-
-  void                        codegenPrototype();
-
-  GenRet                      codegenClassStructType();
-
-  int                         codegenStructure(FILE*       outfile,
-                                               const char* baseoffset);
-
-  int                         codegenFieldStructure(FILE*       outfile,
-                                                    bool        nested,
-                                                    const char* baseOffset);
-
-  // The following two methods are used for types which define initializers
-  bool                        setNextGenericField();
-
-  AggregateType*              getInstantiation(SymExpr* t, int index);
-
-  const char*                 classStructName(bool standalone);
-
-  int                         getMemberGEP(const char* name);
-
-  int                         getFieldPosition(const char* name,
-                                               bool        fatal = true);
-
-  Symbol*                     getField(const char* name, bool fatal = true);
-  Symbol*                     getField(int i);
-
-  // e is as used in PRIM_GET_MEMBER/PRIM_GET_SVEC_MEMBER
-  QualifiedType               getFieldType(Expr* e);
-
-  int                         numFields()                                const;
-
-  bool                        isClass()                                  const;
-  bool                        isRecord()                                 const;
-  bool                        isUnion()                                  const;
-
-  bool                        isGeneric()                                const;
-  void                        markAsGeneric();
-
-  void                        buildConstructors();
-
-  void                        addRootType();
-
-  void                        addClassToHierarchy();
-
-  AggregateTag                aggregateTag;
-
-  InitializerStyle            initializerStyle;
-
-  bool                        initializerResolved;
-
-  AList                       fields;
-
-  // used from parsing, sets dispatchParents
-  AList                       inherits;
-
-  // pointer to an outer class if this is an inner class
-  Symbol*                     outer;
-
-  // Attached only to iterator class/records
-  IteratorInfo*               iteratorInfo;
-
-  // What to delegate to with 'forwarding'
-  AList                       forwardingTo;
-
-  const char*                 doc;
-
-  // Used during code generation for subclass checking,
-  // isa checking. This is the value we store in chpl__cid_XYZ.
-  int                         classId;
-
-private:
-  static ArgSymbol*           createGenericArg(VarSymbol* field);
-  static void                 insertImplicitThis(
-                                            FnSymbol*         fn,
-                                            Vec<const char*>& fieldNamesSet);
-
-private:
-  virtual std::string         docsDirective();
-
-  std::string                 docsSuperClass();
-
-  void                        addDeclaration(DefExpr* defExpr);
-
-  void                        addClassToHierarchy(
-                                          std::set<AggregateType*>& seen);
-
-  AggregateType*              discoverParentAndCheck(Expr* storesName);
-  void                        buildTypeConstructor();
-  void                        buildConstructor();
-  void                        moveConstructorToOuter(FnSymbol* fn);
-
-  std::vector<AggregateType*> instantiations;
-
-  // genericField stores the index of the first generic field in the
-  // AggregateType which does not have a substitution,
-  // but only if the AggregateType defines an initializer.
-  //
-  // If the type has no generic fields without substitutions,
-  // or if setNextGenericField has not been called on the base AggregateType,
-  // this will be set to 0.
-  int                         genericField;
-
-  bool                        mIsGeneric;
-};
+#include "AggregateType.h"
 
 /************************************* | **************************************
 *                                                                             *
@@ -554,26 +414,10 @@ TYPE_EXTERN PrimitiveType* dtTaskID;
 TYPE_EXTERN PrimitiveType* dtSyncVarAuxFields;
 TYPE_EXTERN PrimitiveType* dtSingleVarAuxFields;
 
-// Well-known types
-TYPE_EXTERN AggregateType* dtString;
-TYPE_EXTERN AggregateType* dtArray;
-TYPE_EXTERN AggregateType* dtBaseArr;
-TYPE_EXTERN AggregateType* dtBaseDom;
-TYPE_EXTERN AggregateType* dtDist;
-TYPE_EXTERN AggregateType* dtTuple;
-TYPE_EXTERN AggregateType* dtLocale;
-TYPE_EXTERN AggregateType* dtLocaleID;
-TYPE_EXTERN AggregateType* dtMainArgument;
-
 TYPE_EXTERN PrimitiveType* dtStringC; // the type of a C string (unowned)
 TYPE_EXTERN PrimitiveType* dtStringCopy; // the type of a C string (owned)
 TYPE_EXTERN PrimitiveType* dtCVoidPtr; // the type of a C void* (unowned)
 TYPE_EXTERN PrimitiveType* dtCFnPtr;   // a C function pointer (unowned)
-
-TYPE_EXTERN AggregateType* dtOnBundleRecord;
-TYPE_EXTERN AggregateType* dtTaskBundleRecord;
-
-TYPE_EXTERN AggregateType* dtError;
 
 // base object type (for all classes)
 TYPE_EXTERN Type* dtObject;
